Fix reduction detection bug
To find a reduction, the algorithm is visiting all halide Block, and add a
reduction if it finds a halide For loop doing a reduction initialization,
followed by a halide For loop doing the reduction update.

However, in the previous code, when visiting halide Block to find reductions,
what followed the reduction initialization in the current halide Block was
casted to a halide For, which worked only when there is exactly one halide For
following the initialization.

This meant that reductions where matched only when the update was the last for
loop in the current halide Block during the visit.

In halide2isl, there is by construction no cases where there is an
initialization followed by an invalid update followed by an update. This is
because a reduction was found in tc2halide, which meant that there was an
initialization followed by a valid update. Also, it does not matter if there
is invalid reductions after the found valid reduction.